Settings now fully v5 encrypted. Application detects incorrect passphrase on startup.
The app refuses to attempt to load codes unless it knows the passphrase
is correct. This stops the app from potentially failing to load codes
due to an incorrect passphrase. In such a case, the attempt to load each
code would fail and the user would end up with an empty list regardless
of how many codes s/he actually has. This would be fine while running -
incorrect passphrase = don't load codes. However, on exiting, the
settings would be saved, resulting in am empty list of codes wiping out
the existing codes that were previously in the file. Wiping the list of
codes is not an expected side-effect of getting the passphrase wrong.

The app now behaves as expected - if it detects an incorrect passphrase
it asks for it again without attempting to load any codes. If the user
cancels the passphrase dialogue, the app exits. If the user enters the
correct passphrase, the codes are loaded as normal and the app
continues.