Migrated to storing display plugins in smart pointers rather than raw pointers.
As a pleasant side-effect, this fixes a memory leak when reading codes
from Qr images. The display plugin object for the code read from the QR
was previously created on the heap but never destroyed. Now it is
contained within a std::shared_ptr and therefore is deleted when the
code (which contains the only reference to it) is deleted.

Also added algorithms.h which currently contains a single template class
to compute hashes for std::map (etc.) using qHash() for Qt classes.
Qonvince::QtHash<> can be used as the third template parameter for
std::map containers.